pagination for products is done but search & filter is under process
