+++ package.json
@@ -21,6 +21,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.2.1",
+    "react-infinite-scroll-component": "^6.1.0",
     "react-otp-input": "^3.1.1",
     "react-router-dom": "^6.24.1",
     "zustand": "^4.5.4"
