+++ src/components/navbarpage/searchproduct/CategorizedProduct.css
